Improve horizontal scrolling for small devices
Improve horizontal scrolling visuals for small devices. Boards will often be larger than the screen can handle but we still want to keep it playable with scrolling. Added a new wrapper around the Board to handle the overflow without affecting components outside the board, and added some styles to the scrollbars.