build: help clear up 'nth-check' npm audit warnings
Per react-scripts and Facebook, the 'nth-check' vulnerability warning
is not a problem and not currently planned to be fixed, per
facebook/create-react-app issue number 11647.

Moves 'react-scripts' and testing libraries to devDependencies, to help
with this issue, per facebook/create-react-app issue number 11174.
Audit should now be run with the production flag:
'npm audit --production'