fix: storybook 7 problem with css module 'composes' import from file
Removes 'composes' import from index file for scrollbar CSS in Board
component for the time being, because Storybook 7 has a problem
rendering the component (only within Storybook). Not necessary at this
time for the CSS to be separate anyway, but may revisit in the future
once Storybook 7 is out of beta.